requirements.txt
PyPDF2>=3.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
